+++ tybotRoute/index.js
@@ -284,6 +284,9 @@ async function updateRequestVariables(chatbot, message, projectId, requestId) {
     await chatbot.addParameter(TiledeskChatbotConst.REQ_DEPARTMENT_NAME_KEY, message.request.department.name);
   }
 
+  if (projectId === "641864da99c1fb00131ba495") {
+    console.log("641864da99c1fb00131ba495 > for projectId:", JSON.stringify(message))
+  }
   // for BUG
   // if (chatbot.log) {console.log("message.request.attributes.payload", JSON.stringify(message.request.attributes.payload))}
   if (message && message.request && message.request.attributes && message.request.attributes.payload) {
@@ -292,10 +295,12 @@ async function updateRequestVariables(chatbot, message, projectId, requestId) {
     }
     message.attributes.payload = message.request.attributes.payload
     if (chatbot.log) {console.log("FORCED SET message.attributes.payload:", JSON.stringify(message.attributes.payload))}
+    if (projectId === "641864da99c1fb00131ba495") {console.log("641864da99c1fb00131ba495 > FORCED SET message.attributes.payload:", JSON.stringify(message.attributes.payload))}
   }
 
   if (message.attributes) {
     if (chatbot.log) {console.log("Ok message.attributes", JSON.stringify(message.attributes));}
+    if (projectId === "641864da99c1fb00131ba495") {console.log("641864da99c1fb00131ba495 > Ok message.attributes", JSON.stringify(message.attributes));}
     await chatbot.addParameter(TiledeskChatbotConst.REQ_END_USER_ID_KEY, message.attributes.requester_id);
     await chatbot.addParameter(TiledeskChatbotConst.REQ_END_USER_IP_ADDRESS_KEY, message.attributes.ipAddress);
     if (message.attributes.payload) {
@@ -303,7 +308,7 @@ async function updateRequestVariables(chatbot, message, projectId, requestId) {
         for (const [key, value] of Object.entries(message.attributes.payload)) {
           // const value = all_parameters[key];
           const value_type = typeof value;
-          console.log("importing payload parameter:", key, "value:", value, "type:", value_type)
+          if (projectId === "641864da99c1fb00131ba495") {console.log("641864da99c1fb00131ba495 > importing payload parameter:", key, "value:", value, "type:", value_type);}
           await chatbot.addParameter(key, String(value));
         }
       }